fix: wait language load then switch to target language
