+++ locales/settings/zh-CN.json
@@ -78,7 +78,7 @@
   "feeds.tableHeaders.entryCount": "条目数",
   "feeds.tableHeaders.name": "名称",
   "feeds.tableHeaders.subscriptionCount": "订阅数",
-  "feeds.tableHeaders.tipAmount": "收到的打赏"  ,
+  "feeds.tableHeaders.tipAmount": "收到的打赏",
   "general.app": "应用程序",
   "general.data_persist.description": "在本地保留数据以启用离线访问和本地搜索",
   "general.data_persist.label": "保留数据以供离线使用",
@@ -159,7 +159,7 @@
   "titles.about": "关于",
   "titles.actions": "自动化",
   "titles.appearance": "外观",
-  "titles.general": "常规",
+  "titles.general": "通用",
   "titles.integration": "集成",
   "titles.invitations": "邀请",
   "titles.power": "Power",
